// offchain/src/generated/openbounty/types/prizeTier.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getAddressDecoder,
  getAddressEncoder,
  getArrayDecoder,
  getArrayEncoder,
  getBooleanDecoder,
  getBooleanEncoder,
  getOptionDecoder,
  getOptionEncoder,
  getStructDecoder,
  getStructEncoder,
  getU64Decoder,
  getU64Encoder,
  type Address,
  type Codec,
  type Decoder,
  type Encoder,
  type Option,
  type OptionOrNullable,
} from "@solana/kit";
import { getVoteDecoder, getVoteEncoder, type Vote, type VoteArgs } from ".";

export type PrizeTier = {
  /** Lamports paid to this tier's winner. */
  amount: bigint;
  /** Set by the program when a candidate reaches `vote_threshold`. */
  winner: Option<Address>;
  claimed: boolean;
  /** Set when the organizer reclaims this tier's amount after the deadline. */
  refunded: boolean;
  votes: Array<Vote>;
};

export type PrizeTierArgs = {
  /** Lamports paid to this tier's winner. */
  amount: number | bigint;
  /** Set by the program when a candidate reaches `vote_threshold`. */
  winner: OptionOrNullable<Address>;
  claimed: boolean;
  /** Set when the organizer reclaims this tier's amount after the deadline. */
  refunded: boolean;
  votes: Array<VoteArgs>;
};

export function getPrizeTierEncoder(): Encoder<PrizeTierArgs> {
  return getStructEncoder([
    ["amount", getU64Encoder()],
    ["winner", getOptionEncoder(getAddressEncoder())],
    ["claimed", getBooleanEncoder()],
    ["refunded", getBooleanEncoder()],
    ["votes", getArrayEncoder(getVoteEncoder())],
  ]);
}

export function getPrizeTierDecoder(): Decoder<PrizeTier> {
  return getStructDecoder([
    ["amount", getU64Decoder()],
    ["winner", getOptionDecoder(getAddressDecoder())],
    ["claimed", getBooleanDecoder()],
    ["refunded", getBooleanDecoder()],
    ["votes", getArrayDecoder(getVoteDecoder())],
  ]);
}

export function getPrizeTierCodec(): Codec<PrizeTierArgs, PrizeTier> {
  return combineCodec(getPrizeTierEncoder(), getPrizeTierDecoder());
}